Gnome Search Provider for Ripasso
Implements a GNOME search provider for Ripasso. This is a very
convenient way to obtain passwords from pass when using the GNOME shell.

I totally understand if this isn't the kind of addition you would like
to see in this repo, especially if you don't use GNOME. So I have no
problem spinning it off into it's own project. It's only about 150 LoC.
But since there is already a GTK project in here, I figured why not give
it a shot.

There will need to be some additions made for the configuration files. I
know there is at least an arch linux package, and I'd be happy to
contribute to that.

Thanks for your work on Ripasso!