Relax requirement for 128 bit totp secrets
totp-rs is strictly RFC6238 compliant. This is a good thing, but the
reality is many sites/apps are still using 80 bit secrets for TOTP.
These include Github, Discord, Paypal, among others. The author of
totp-rs added a function `from_url_unchecked` to address this in this
issue: https://github.com/constantoine/totp-rs/issues/46. I suggest we
use it here so that ripasso can be used practically for totp.